src/third_party/WebKit/LayoutTests/fast/dom/shadow/resize-in-shadow-dom-expected.txt

CONSOLE WARNING: line 22: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS box.offsetWidth is 114
PASS box.offsetHeight is 114
PASS successfullyParsed is true

TEST COMPLETE
